[ASV-1415] Refactor: Moved Editorial related files to a specific editorial package
