+++ app/src/main/java/cm/aptoide/pt/editorial/EditorialEvent.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.app.view;
+package cm.aptoide.pt.editorial;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import java.util.List;
  * Created by D01 on 19/09/2018.
  */
 
-class EditorialEvent {
+public class EditorialEvent {
 
   private final Type clickType;
   private final int firstVisibleItemPosition;
