+++ app/src/main/java/cm/aptoide/pt/editorial/EditorialFragment.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.app.view;
+package cm.aptoide.pt.editorial;
 
 import android.animation.Animator;
 import android.content.res.Resources;
