+++ app/src/main/java/cm/aptoide/pt/editorial/EditorialItemsViewHolder.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.app.view;
+package cm.aptoide.pt.editorial;
 
 import android.graphics.Rect;
 import android.support.v7.widget.LinearLayoutManager;
