+++ app/src/main/java/cm/aptoide/pt/editorial/EditorialManager.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.app.view;
+package cm.aptoide.pt.editorial;
 
 import cm.aptoide.analytics.AnalyticsManager;
 import cm.aptoide.pt.app.DownloadModel;
