+++ app/src/main/java/cm/aptoide/pt/editorial/EditorialRepository.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.app.view;
+package cm.aptoide.pt.editorial;
 
 import rx.Single;
