+++ app/src/main/java/cm/aptoide/pt/editorial/EditorialService.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.app.view;
+package cm.aptoide.pt.editorial;
 
 import android.content.SharedPreferences;
 import cm.aptoide.pt.dataprovider.exception.NoNetworkConnectionException;
