+++ app/src/main/java/cm/aptoide/pt/editorial/EditorialView.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.app.view;
+package cm.aptoide.pt.editorial;
 
 import android.support.v7.graphics.Palette;
 import cm.aptoide.pt.app.DownloadModel;
