+++ app/src/main/java/cm/aptoide/pt/editorial/MediaBundleAdapter.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.app.view;
+package cm.aptoide.pt.editorial;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
