+++ app/src/main/java/cm/aptoide/pt/view/FragmentModule.java
@@ -48,13 +48,6 @@ import cm.aptoide.pt.app.view.AppViewFragment.BundleKeys;
 import cm.aptoide.pt.app.view.AppViewNavigator;
 import cm.aptoide.pt.app.view.AppViewPresenter;
 import cm.aptoide.pt.app.view.AppViewView;
-import cm.aptoide.pt.app.view.EditorialAnalytics;
-import cm.aptoide.pt.app.view.EditorialManager;
-import cm.aptoide.pt.app.view.EditorialNavigator;
-import cm.aptoide.pt.app.view.EditorialPresenter;
-import cm.aptoide.pt.app.view.EditorialRepository;
-import cm.aptoide.pt.app.view.EditorialService;
-import cm.aptoide.pt.app.view.EditorialView;
 import cm.aptoide.pt.app.view.MoreBundleManager;
 import cm.aptoide.pt.app.view.MoreBundlePresenter;
 import cm.aptoide.pt.app.view.MoreBundleView;
@@ -68,6 +61,13 @@ import cm.aptoide.pt.dataprovider.model.v7.Type;
 import cm.aptoide.pt.dataprovider.ws.BodyInterceptor;
 import cm.aptoide.pt.dataprovider.ws.v7.BaseBody;
 import cm.aptoide.pt.download.DownloadFactory;
+import cm.aptoide.pt.editorial.EditorialAnalytics;
+import cm.aptoide.pt.editorial.EditorialManager;
+import cm.aptoide.pt.editorial.EditorialNavigator;
+import cm.aptoide.pt.editorial.EditorialPresenter;
+import cm.aptoide.pt.editorial.EditorialRepository;
+import cm.aptoide.pt.editorial.EditorialService;
+import cm.aptoide.pt.editorial.EditorialView;
 import cm.aptoide.pt.home.AdMapper;
 import cm.aptoide.pt.home.AptoideBottomNavigator;
 import cm.aptoide.pt.home.BannerRepository;
