+++ app/src/test/java/cm/aptoide/pt/app/view/EditorialPresenterTest.java
@@ -6,6 +6,15 @@ import cm.aptoide.pt.actions.PermissionManager;
 import cm.aptoide.pt.actions.PermissionService;
 import cm.aptoide.pt.app.DownloadModel;
 import cm.aptoide.pt.crashreports.CrashReport;
+import cm.aptoide.pt.editorial.EditorialAnalytics;
+import cm.aptoide.pt.editorial.EditorialContent;
+import cm.aptoide.pt.editorial.EditorialEvent;
+import cm.aptoide.pt.editorial.EditorialFragment;
+import cm.aptoide.pt.editorial.EditorialManager;
+import cm.aptoide.pt.editorial.EditorialNavigator;
+import cm.aptoide.pt.editorial.EditorialPresenter;
+import cm.aptoide.pt.editorial.EditorialViewModel;
+import cm.aptoide.pt.editorial.ScrollEvent;
 import cm.aptoide.pt.presenter.View;
 import java.util.ArrayList;
 import java.util.Collections;
